embed test to a related for throughput
