+++ tests/std/test.lst
@@ -155,7 +155,6 @@ tests\Dev11_1140665_unique_ptr_array_conversions
 tests\Dev11_1150223_shared_mutex
 tests\Dev11_1158803_regex_thread_safety
 tests\Dev11_1180290_filesystem_error_code
-tests\GH_000258_sometimes_avoid_locking
 tests\GH_000431_copy_move_family
 tests\GH_000431_equal_family
 tests\GH_000431_equal_memcmp_is_safe
