+++ tests/std/tests/Dev10_851347_weak_ptr_virtual_inheritance/test.cpp
@@ -2,8 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include <assert.h>
+#include <atomic>
 #include <memory>
 #include <utility>
+#include <thread>
 using namespace std;
 
 // Also test GH-1102 "<memory>: weak_ptr conversions don't preserve control blocks for expired objects"
@@ -76,6 +78,53 @@ struct D : B, C {
     int d{40};
 };
 
+void test_gh_000258_sometimes_avoid_locking() {
+    struct base1 {
+        int i = 0;
+    };
+
+    struct base2 {
+        int j = 0;
+    };
+
+    struct base3 {
+        int k = 0;
+    };
+
+    struct derived : virtual base1, virtual base2, base3 {};
+
+    for (int i = 0; i < 10; ++i) {
+        // not make_shared -- with make_shared the test would not catch errors
+        shared_ptr<derived> d{new derived{}};
+        weak_ptr<derived> wd{d};
+        atomic<bool> work{true};
+        thread thd([&] {
+            d.reset();
+            this_thread::yield(); // make crash on incorrect optimization even more likely
+            work = false;
+        });
+
+        if ((i % 2) == 0) {
+            while (work) {
+                // likely to crash if optimized for a case we shouldn't
+                weak_ptr<base1> wb1{wd};
+                weak_ptr<base2> wb2{wd};
+                weak_ptr<base3> wb3{wd};
+            }
+        } else {
+            while (work) {
+                // likely to crash if optimized for a case we shouldn't
+                weak_ptr<base1> wb1{weak_ptr<derived>{wd}};
+                weak_ptr<base2> wb2{weak_ptr<derived>{wd}};
+                weak_ptr<base3> wb3{weak_ptr<derived>{wd}};
+            }
+        }
+
+        thd.join();
+    }
+}
+
+
 int main() {
     test_owner_equal();
 
@@ -109,4 +158,6 @@ int main() {
     wpa4 = move(wpd_two);
     assert(wpa4.expired());
     assert(owner_equal(wpa4, wpd_zero));
+
+    test_gh_000258_sometimes_avoid_locking();
 }
